services/rag/requirements.txt
pymilvus
sentence-transformers
langchain
langchain-community
ollama
python-dotenv
